refactor(entities): перенести CharacterBuilder из tests в движок
CharacterBuilder используется существующими модульными тестами для
программного создания CharData с заданным классом, уровнем и аффектами.
Автономному симулятору баланса (issue #2967) нужен такой же построитель.
Чтобы не дублировать код, переносим класс в src/engine/entities/, меняем
namespace test_utils -> entities. В tests/char.utilities.hpp оставляем
тонкую прослойку с typedef-ом в test_utils, чтобы существующие тесты не
пришлось править.

Логика без изменений (расширение API — отдельным коммитом). Все 388
существующих тестов плюс 4 теста FileEventSink и 3 теста RandomSeed
проходят (392/392).
