+++ CMakeLists.txt
@@ -37,6 +37,7 @@ set(SOURCES
 		src/utils/cache.cpp
 		src/gameplay/mechanics/celebrates.cpp
 		src/engine/entities/char_data.cpp
+		src/engine/entities/character_builder.cpp
 		src/engine/entities/char_player.cpp
 		src/gameplay/classes/pc_classes.cpp
 		src/gameplay/classes/classes_spell_slots.cpp
@@ -566,6 +567,7 @@ set(HEADERS
 		src/utils/cache.h
 		src/gameplay/mechanics/celebrates.h
 		src/engine/entities/char_data.h
+		src/engine/entities/character_builder.h
 		src/engine/core/utils_char_obj.inl
 		src/engine/entities/char_player.h
 		src/gameplay/classes/pc_classes.h
