+++ src/engine/entities/character_builder.h
@@ -1,0 +1,42 @@
+#ifndef BYLINS_CHARACTER_BUILDER_H
+#define BYLINS_CHARACTER_BUILDER_H
+
+#include "char_data.h"
+
+namespace entities {
+
+class CharacterBuilder {
+public:
+	using character_t = CharData;
+	using result_t = character_t::shared_ptr;
+
+	void create_new();
+	void create_new_with_class(const short player_class);
+	void create_character_with_one_removable_affect();
+	void create_character_with_two_removable_affects();
+	void create_character_with_two_removable_and_two_not_removable_affects();
+
+	void add_poison();
+	void add_sleep();
+	void add_detect_invis();
+	void add_detect_align();
+	void set_level(const int level);
+	void set_class(const short player_class);
+
+	void make_group(CharacterBuilder& character_builder);
+
+	result_t get() const { return m_result; }
+
+private:
+	void check_character_existance() const;
+
+	static void check_character_existance(result_t character);
+
+	result_t m_result;
+};
+
+}  // namespace entities
+
+#endif  // BYLINS_CHARACTER_BUILDER_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
