+++ tests/char.utilities.hpp
@@ -1,40 +1,12 @@
 #ifndef __CHAR__UTILITIES_HPP__
 #define __CHAR__UTILITIES_HPP__
 
-#include "engine/entities/char_data.h"
+// Backwards-compat shim: CharacterBuilder lives in src/engine/entities/ now,
+// so it can be used by both tests and the headless balance simulator.
+#include "engine/entities/character_builder.h"
 
-namespace test_utils
-{
-	class CharacterBuilder
-	{
-	public:
-		using character_t = CharData;
-		using result_t = character_t::shared_ptr;
-
-		void create_new();
-		void create_new_with_class(const short player_class);
-		void create_character_with_one_removable_affect();
-		void create_character_with_two_removable_affects();
-		void create_character_with_two_removable_and_two_not_removable_affects();
-
-		void add_poison();
-		void add_sleep();
-		void add_detect_invis();
-		void add_detect_align();
-		void set_level(const int level);
-		void set_class(const short player_class);
-
-		void make_group(CharacterBuilder& character_builder);
-
-		result_t get() const { return m_result; }
-
-	private:
-		void check_character_existance() const;
-
-		static void check_character_existance(result_t character);
-
-		result_t m_result;
-	};
+namespace test_utils {
+	using CharacterBuilder = entities::CharacterBuilder;
 }
 
 #endif // __CHAR__UTILITIES_HPP__
